+++ app/src/androidTest/java/org/simple/clinic/rules/ServerAuthenticationRule.kt
@@ -111,9 +111,7 @@ class ServerAuthenticationRule : TestRule {
   }
 
   private fun ensureFacilities() {
-    val result = facilitySync
-        .pullWithResult()
-        .blockingGet()
+    val result = facilitySync.pullWithResult()
 
     assertThat(result).isEqualTo(FacilityPullResult.Success)
   }